Add MapApiGroup and validation
