+++ examples/AspNetCore/WebApi/MinimalOpenApiExample/Program.cs
@@ -47,217 +47,220 @@ services.AddSwaggerGen( options => options.OperationFilter<SwaggerDefaultValues>
 
 // Configure the HTTP request pipeline.
 var app = builder.Build();
-var orders = app.MapGroup( "/api/orders" ).WithApiVersionSet( "Orders" );
-var people = app.MapGroup( "/api/v{version:apiVersion}/people" ).WithApiVersionSet( "People" );
+var orders = app.MapApiGroup( "Orders" );
+var people = app.MapApiGroup( "People" );
 
 // 1.0
-orders.MapGet( "/{id:int}", ( int id ) => new OrderV1() { Id = id, Customer = "John Doe" } )
-      .Produces<OrderV1>()
-      .Produces( 404 )
-      .HasDeprecatedApiVersion( 0.9 )
-      .HasApiVersion( 1.0 );
+var o1 = orders.MapGroup( "/api/orders" )
+               .HasDeprecatedApiVersion( 0.9 )
+               .HasApiVersion( 1.0 );
 
-orders.MapPost( "/", ( HttpRequest request, OrderV1 order ) =>
-       {
-           order.Id = 42;
-           var scheme = request.Scheme;
-           var host = request.Host;
-           var location = new Uri( $"{scheme}{Uri.SchemeDelimiter}{host}/api/orders/{order.Id}" );
-           return Results.Created( location, order );
-       } )
-      .Accepts<OrderV1>( "application/json" )
-      .Produces<OrderV1>( 201 )
-      .Produces( 400 )
-      .HasApiVersion( 1.0 );
+o1.MapGet( "/{id:int}", ( int id ) => new OrderV1() { Id = id, Customer = "John Doe" } )
+  .Produces<OrderV1>()
+  .Produces( 404 );
 
-orders.MapMethods( "/{id:int}", new[] { HttpMethod.Patch.Method }, ( int id, OrderV1 order ) => Results.NoContent() )
-      .Accepts<OrderV1>( "application/json" )
-      .Produces( 204 )
-      .Produces( 400 )
-      .Produces( 404 )
-      .HasApiVersion( 1.0 );
+o1.MapPost( "/", ( HttpRequest request, OrderV1 order ) =>
+   {
+       order.Id = 42;
+       var scheme = request.Scheme;
+       var host = request.Host;
+       var location = new Uri( $"{scheme}{Uri.SchemeDelimiter}{host}/api/orders/{order.Id}" );
+       return Results.Created( location, order );
+   } )
+  .Accepts<OrderV1>( "application/json" )
+  .Produces<OrderV1>( 201 )
+  .Produces( 400 )
+  .MapToApiVersion( 1.0 );
+
+o1.MapMethods( "/{id:int}", new[] { HttpMethod.Patch.Method }, ( int id, OrderV1 order ) => Results.NoContent() )
+  .Accepts<OrderV1>( "application/json" )
+  .Produces( 204 )
+  .Produces( 400 )
+  .Produces( 404 )
+  .MapToApiVersion( 1.0 );
 
 // 2.0
-orders.MapGet( "/", () =>
-       new OrderV2[]
-       {
-           new(){ Id = 1, Customer = "John Doe" },
-           new(){ Id = 2, Customer = "Bob Smith" },
-           new(){ Id = 3, Customer = "Jane Doe", EffectiveDate = DateTimeOffset.UtcNow.AddDays( 7d ) },
-       } )
-      .Produces<IEnumerable<OrderV2>>()
-      .Produces( 404 )
-      .HasApiVersion( 2.0 );
+var o2 = orders.MapGroup( "/api/orders" )
+               .HasApiVersion( 2.0 );
 
-orders.MapGet( "/{id:int}", ( int id ) => new OrderV2() { Id = id, Customer = "John Doe" } )
-      .Produces<OrderV2>()
-      .Produces( 404 )
-      .HasApiVersion( 2.0 );
+o2.MapGet( "/", () =>
+   new OrderV2[]
+   {
+       new(){ Id = 1, Customer = "John Doe" },
+       new(){ Id = 2, Customer = "Bob Smith" },
+       new(){ Id = 3, Customer = "Jane Doe", EffectiveDate = DateTimeOffset.UtcNow.AddDays( 7d ) },
+   } )
+  .Produces<IEnumerable<OrderV2>>()
+  .Produces( 404 );
 
-orders.MapPost( "/", ( HttpRequest request, OrderV2 order ) =>
-       {
-           order.Id = 42;
-           var scheme = request.Scheme;
-           var host = request.Host;
-           var location = new Uri( $"{scheme}{Uri.SchemeDelimiter}{host}/api/orders/{order.Id}" );
-           return Results.Created( location, order );
-       } )
-      .Accepts<OrderV2>( "application/json" )
-      .Produces<OrderV2>( 201 )
-      .Produces( 400 )
-      .HasApiVersion( 2.0 );
+o2.MapGet( "/{id:int}", ( int id ) => new OrderV2() { Id = id, Customer = "John Doe" } )
+  .Produces<OrderV2>()
+  .Produces( 404 );
 
-orders.MapMethods( "/{id:int}", new[] { HttpMethod.Patch.Method }, ( int id, OrderV2 order ) => Results.NoContent() )
-      .Accepts<OrderV2>( "application/json" )
-      .Produces( 204 )
-      .Produces( 400 )
-      .Produces( 404 )
-      .HasApiVersion( 2.0 );
+o2.MapPost( "/", ( HttpRequest request, OrderV2 order ) =>
+   {
+       order.Id = 42;
+       var scheme = request.Scheme;
+       var host = request.Host;
+       var location = new Uri( $"{scheme}{Uri.SchemeDelimiter}{host}/api/orders/{order.Id}" );
+       return Results.Created( location, order );
+   } )
+  .Accepts<OrderV2>( "application/json" )
+  .Produces<OrderV2>( 201 )
+  .Produces( 400 );
+
+o2.MapMethods( "/{id:int}", new[] { HttpMethod.Patch.Method }, ( int id, OrderV2 order ) => Results.NoContent() )
+  .Accepts<OrderV2>( "application/json" )
+  .Produces( 204 )
+  .Produces( 400 )
+  .Produces( 404 );
 
 // 3.0
-orders.MapGet( "/", () =>
-       new OrderV3[]
-       {
-           new(){ Id = 1, Customer = "John Doe" },
-           new(){ Id = 2, Customer = "Bob Smith" },
-           new(){ Id = 3, Customer = "Jane Doe", EffectiveDate = DateTimeOffset.UtcNow.AddDays( 7d ) },
-       } )
-      .Produces<IEnumerable<OrderV3>>()
-      .HasApiVersion( 3.0 );
+var o3 = orders.MapGroup( "/api/orders" )
+               .HasApiVersion( 3.0 );
 
-orders.MapGet( "/{id:int}", ( int id ) => new OrderV3() { Id = id, Customer = "John Doe" } )
-      .Produces<OrderV3>()
-      .Produces( 404 )
-      .HasApiVersion( 3.0 );
+o3.MapGet( "/", () =>
+   new OrderV3[]
+   {
+       new(){ Id = 1, Customer = "John Doe" },
+       new(){ Id = 2, Customer = "Bob Smith" },
+       new(){ Id = 3, Customer = "Jane Doe", EffectiveDate = DateTimeOffset.UtcNow.AddDays( 7d ) },
+   } )
+  .Produces<IEnumerable<OrderV3>>();
 
-orders.MapPost( "/", ( HttpRequest request, OrderV3 order ) =>
-       {
-           order.Id = 42;
-           var scheme = request.Scheme;
-           var host = request.Host;
-           var location = new Uri( $"{scheme}{Uri.SchemeDelimiter}{host}/api/orders/{order.Id}" );
-           return Results.Created( location, order );
-       } )
-      .Accepts<OrderV3>( "application/json" )
-      .Produces<OrderV3>( 201 )
-      .Produces( 400 )
-      .HasApiVersion( 3.0 );
+o3.MapGet( "/{id:int}", ( int id ) => new OrderV3() { Id = id, Customer = "John Doe" } )
+  .Produces<OrderV3>()
+  .Produces( 404 );
 
-orders.MapDelete( "/{id:int}", ( int id ) => Results.NoContent() )
-      .Produces( 204 )
-      .HasApiVersion( 3.0 );
+o3.MapPost( "/", ( HttpRequest request, OrderV3 order ) =>
+   {
+       order.Id = 42;
+       var scheme = request.Scheme;
+       var host = request.Host;
+       var location = new Uri( $"{scheme}{Uri.SchemeDelimiter}{host}/api/orders/{order.Id}" );
+       return Results.Created( location, order );
+   } )
+  .Accepts<OrderV3>( "application/json" )
+  .Produces<OrderV3>( 201 )
+  .Produces( 400 );
+
+o3.MapDelete( "/{id:int}", ( int id ) => Results.NoContent() )
+  .Produces( 204 );
 
 // 1.0
-people.MapGet( "/{id:int}", ( int id ) =>
-       new PersonV1()
-       {
-           Id = id,
-           FirstName = "John",
-           LastName = "Doe",
-       } )
-      .Produces<PersonV1>()
-      .Produces( 404 )
-      .HasDeprecatedApiVersion( 0.9 )
-      .HasApiVersion( 1.0 );
+var p1 = people.MapGroup( "/api/v{version:apiVersion}/people" )
+               .HasDeprecatedApiVersion( 0.9 )
+               .HasApiVersion( 1.0 );
+
+p1.MapGet( "/{id:int}", ( int id ) =>
+   new PersonV1()
+   {
+       Id = id,
+       FirstName = "John",
+       LastName = "Doe",
+   } )
+  .Produces<PersonV1>()
+  .Produces( 404 );
 
 // 2.0
-people.MapGet( "/", () =>
-       new PersonV2[]
-       {
-           new()
-           {
-               Id = 1,
-               FirstName = "John",
-               LastName = "Doe",
-               Email = "john.doe@somewhere.com",
-           },
-           new()
-           {
-               Id = 2,
-               FirstName = "Bob",
-               LastName = "Smith",
-               Email = "bob.smith@somewhere.com",
-           },
-           new()
-           {
-               Id = 3,
-               FirstName = "Jane",
-               LastName = "Doe",
-               Email = "jane.doe@somewhere.com",
-           },
-       } )
-      .Produces<IEnumerable<PersonV2>>()
-      .HasApiVersion( 2.0 );
+var p2 = people.MapGroup( "/api/v{version:apiVersion}/people" )
+               .HasApiVersion( 2.0 );
 
-people.MapGet( "/{id:int}", ( int id ) =>
-       new PersonV2()
+p2.MapGet( "/", () =>
+   new PersonV2[]
+   {
+       new()
        {
-           Id = id,
+           Id = 1,
            FirstName = "John",
            LastName = "Doe",
            Email = "john.doe@somewhere.com",
-       } )
-      .Produces<PersonV2>()
-      .Produces( 404 )
-      .HasApiVersion( 2.0 );
+       },
+       new()
+       {
+           Id = 2,
+           FirstName = "Bob",
+           LastName = "Smith",
+           Email = "bob.smith@somewhere.com",
+       },
+       new()
+       {
+           Id = 3,
+           FirstName = "Jane",
+           LastName = "Doe",
+           Email = "jane.doe@somewhere.com",
+       },
+   } )
+  .Produces<IEnumerable<PersonV2>>();
+
+p2.MapGet( "/{id:int}", ( int id ) =>
+   new PersonV2()
+   {
+       Id = id,
+       FirstName = "John",
+       LastName = "Doe",
+       Email = "john.doe@somewhere.com",
+   } )
+  .Produces<PersonV2>()
+  .Produces( 404 );
 
 // 3.0
-people.MapGet( "/", () =>
-       new PersonV3[]
-       {
-           new()
-           {
-               Id = 1,
-               FirstName = "John",
-               LastName = "Doe",
-               Email = "john.doe@somewhere.com",
-               Phone = "555-987-1234",
-           },
-           new()
-           {
-               Id = 2,
-               FirstName = "Bob",
-               LastName = "Smith",
-               Email = "bob.smith@somewhere.com",
-               Phone = "555-654-4321",
-           },
-           new()
-           {
-               Id = 3,
-               FirstName = "Jane",
-               LastName = "Doe",
-               Email = "jane.doe@somewhere.com",
-               Phone = "555-789-3456",
-           },
-       } )
-      .Produces<IEnumerable<PersonV3>>()
-      .HasApiVersion( 3.0 );
+var p3 = people.MapGroup( "/api/v{version:apiVersion}/people" )
+               .HasApiVersion( 3.0 );
 
-people.MapGet( "/{id:int}", ( int id ) =>
-       new PersonV3()
+p3.MapGet( "/", () =>
+   new PersonV3[]
+   {
+       new()
        {
-           Id = id,
+           Id = 1,
            FirstName = "John",
            LastName = "Doe",
            Email = "john.doe@somewhere.com",
            Phone = "555-987-1234",
-       } )
-      .Produces<PersonV3>()
-      .Produces( 404 )
-      .HasApiVersion( 3.0 );
-
-people.MapPost( "/", ( HttpRequest request, ApiVersion version, PersonV3 person ) =>
+       },
+       new()
        {
-           person.Id = 42;
-           var scheme = request.Scheme;
-           var host = request.Host;
-           var location = new Uri( $"{scheme}{Uri.SchemeDelimiter}{host}/v{version}/api/people/{person.Id}" );
-           return Results.Created( location, person );
-       } )
-      .Accepts<PersonV3>( "application/json" )
-      .Produces<PersonV3>( 201 )
-      .Produces( 400 )
-      .HasApiVersion( 3.0 );
+           Id = 2,
+           FirstName = "Bob",
+           LastName = "Smith",
+           Email = "bob.smith@somewhere.com",
+           Phone = "555-654-4321",
+       },
+       new()
+       {
+           Id = 3,
+           FirstName = "Jane",
+           LastName = "Doe",
+           Email = "jane.doe@somewhere.com",
+           Phone = "555-789-3456",
+       },
+   } )
+  .Produces<IEnumerable<PersonV3>>();
+
+p3.MapGet( "/{id:int}", ( int id ) =>
+   new PersonV3()
+   {
+       Id = id,
+       FirstName = "John",
+       LastName = "Doe",
+       Email = "john.doe@somewhere.com",
+       Phone = "555-987-1234",
+   } )
+  .Produces<PersonV3>()
+  .Produces( 404 );
+
+p3.MapPost( "/", ( HttpRequest request, ApiVersion version, PersonV3 person ) =>
+   {
+       person.Id = 42;
+       var scheme = request.Scheme;
+       var host = request.Host;
+       var location = new Uri( $"{scheme}{Uri.SchemeDelimiter}{host}/v{version}/api/people/{person.Id}" );
+       return Results.Created( location, person );
+   } )
+  .Accepts<PersonV3>( "application/json" )
+  .Produces<PersonV3>( 201 )
+  .Produces( 400 );
 
 app.UseSwagger();
 app.UseSwaggerUI(
